+++ R/sig_auto_extract.R
@@ -108,6 +108,13 @@ sig_auto_extract <- function(nmf_matrix = NULL,
   if (!recover) {
     nmf_matrix <- t(nmf_matrix) # rows for mutation types and columns for samples
 
+    ii <- colSums(nmf_matrix) < 0.01
+    if (any(ii)) {
+      message("The follow samples dropped due to null catalogue:\n\t",
+              paste0(colnames(nmf_matrix)[ii], collapse = ", "))
+      nmf_matrix <- nmf_matrix[, !ii, drop = FALSE]
+    }
+
     oplan <- future::plan()
     future::plan("multiprocess", workers = cores)
     on.exit(future::plan(oplan), add = TRUE)
@@ -177,8 +184,6 @@ sig_auto_extract <- function(nmf_matrix = NULL,
         sig = W_cn,
         method = "QP",
         mode = "copynumber")
-      best_solution$Exposure.norm <- apply(best_solution$Exposure, 2,
-                                           function(x) x / sum(x, na.rm = TRUE))
     } else {
       ## Call LCD
       best_solution$Exposure <- sig_fit(
@@ -186,8 +191,43 @@ sig_auto_extract <- function(nmf_matrix = NULL,
         sig = apply(best_solution$Signature,
                     2, function(x) x / sum(x)),
         method = "QP")
-      best_solution$Exposure.norm <- apply(best_solution$Exposure, 2,
-                                           function(x) x / sum(x, na.rm = TRUE))
+    }
+
+    Exposure <- best_solution$Exposure
+    # Handle hyper mutant samples
+    hyper_index <- grepl("_\\[hyper\\]_", colnames(Exposure))
+    if (sum(hyper_index) > 0) {
+      H.hyper <- Exposure[, hyper_index, drop = FALSE]
+      H.nonhyper <- Exposure[, !hyper_index, drop = FALSE]
+      sample.hyper <- sapply(
+        colnames(H.hyper),
+        function(x) strsplit(x, "_\\[hyper\\]_")[[1]][[1]]
+      )
+      unique.hyper <- unique(sample.hyper)
+      n.hyper <- length(unique.hyper)
+      x.hyper <- array(0, dim = c(nrow(H.hyper), n.hyper))
+      for (i in 1:n.hyper) {
+        x.hyper[, i] <- rowSums(H.hyper[, sample.hyper %in% unique.hyper[i], drop = FALSE])
+      }
+      colnames(x.hyper) <- unique.hyper
+      rownames(x.hyper) <- rownames(Exposure)
+      Exposure <- cbind(H.nonhyper, x.hyper)
+      best_solution$Exposure <- Exposure
+    }
+
+    best_solution$Exposure.norm <- apply(best_solution$Exposure, 2,
+                                         function(x) x / sum(x, na.rm = TRUE))
+    # When only one signature
+    if (!is.matrix(best_solution$Exposure.norm)) {
+      best_solution$Exposure.norm <- matrix(best_solution$Exposure.norm,
+                                            nrow = 1,
+                                            dimnames = list(NULL, names(best_solution$Exposure.norm)))
+    }
+    ## Scale the result
+    if (any(has_cn)) {
+      best_solution$Signature <- best_solution$Signature.norm * sum(nmf_matrix, na.rm = TRUE)
+    } else {
+      best_solution$Signature <- best_solution$Signature.norm * sum(best_solution$Exposure, na.rm = TRUE)
     }
   }