+++ R/sigprofiler.R
@@ -128,8 +128,16 @@ sigprofiler_extract <- function(nmf_matrix, output, range = 2:5, nrun = 10L,
 
   # print(sys$path)
 
+  nmf_matrix <- t(nmf_matrix)
+
+  ii <- colSums(nmf_matrix) < 0.01
+  if (any(ii)) {
+    message("The follow samples dropped due to null catalogue:\n\t",
+            paste0(colnames(nmf_matrix)[ii], collapse = ", "))
+    nmf_matrix <- nmf_matrix[, !ii, drop = FALSE]
+  }
+
   in_df <- nmf_matrix %>%
-    t() %>%
     as.data.frame() %>%
     tibble::rownames_to_column("MutationType")